image_server/requirements.txt
asgiref~=3.2
jsonfield
sqlparse
certifi
decorator
Django
django-cors-headers
gunicorn
hexbytes
oauthlib
Pillow
psycopg2
ptyprocess
python-dateutil
pytz
redis
wcwidth
websockets
django-environ
psycopg2-binary
PyJWT
djangorestframework
tzdata